developed new forAll that works with two generators at once
In case we have both a generator for type A and a generator for
predicates A => Boolean, the new forAll implements the following
composition of two forAlls:

forAll(agen: Gen[A])(a => forAll(fgen: Gen[A => Boolean])(f => f(a)))

This looks like a flatMap.  It might be worthwhile to make this explicit
by implementing flatMap for Prop type.