minor changes.
(Noticed some weird behavior of the Scala worksheet lately.  When I save
it, Eclipse tries to compile the worksheet and print results, but gets
stuck in the middle of the worksheet, and the * (unsaved file) symbol
goes away for only a few seconds, but then comes back.  However, when I
commit changes to the worksheet with git, Eclipse asks if I want to save
unsaved changes.  When I say yes, then the worksheet gets processed
properly, will the correct output, and the file is save (the * symbol
goes away)... weird.  Maybe next time I'll try saving the worksheet
using the File-> Save command (instead of Ctrl-x Ctrl-s).